Abstract classes to interface
