--- sashimi/hardware/external_trigger/__init__.py
@@ -1,9 +1,0 @@
-from typing import Optional
-
-
-class AbstractComm:
-    def __init__(self, address):
-        self.address = address
-
-    def trigger_and_receive_duration(self, config) -> Optional[float]:
-        return None
