+++ sashimi/hardware/external_trigger/interface.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+
+class AbstractComm:
+    def __init__(self, address):
+        self.address = address
+
+    def trigger_and_receive_duration(self, config) -> Optional[float]:
+        return None
